+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- SLI pluggins support.
+- SLI plugins support.
 
 ## [v0.2.0] - 2021-05-24