+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [v0.3.0] - 2021-06-10
+
 ### Added
 
 - SLI plugins support.
@@ -52,6 +54,7 @@
 - Support raw query based SLI.
 - Kubernetes (prometheus-operator) CRD generation support.
 
-[unreleased]: https://github.com/slok/sloth/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/slok/sloth/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/slok/sloth/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/slok/sloth/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/slok/sloth/releases/tag/v0.1.0